+++ src/apps/depthBuffer/main.cpp
@@ -311,7 +311,7 @@ void keyCallback(int key, int, int action, int) {
                 break;
             case key::F:
                 if (action == action::Press) {
-                    for (size_t i = 0; i < gEngine->getNumberOfWindows(); i++) {
+                    for (int i = 0; i < gEngine->getNumberOfWindows(); i++) {
                         gEngine->getWindow(i).setUseFXAA(
                             !gEngine->getWindow(i).useFXAA()
                         );
